116 parameters header: changed speed to a long

+++ include/wus_pulse_out.h
@@ -37,6 +37,6 @@ void initPulseOut();
  *
  * \param speed The speed.
  */
-void setPulseSpeed(int speed);
+void setPulseSpeed(long speed);
 
 #endif
